+++ lib/common/utils.dart
@@ -1,0 +1,3 @@
+part of '../d_chart.dart';
+
+double radianFromDegree(double degree) => degree * math.pi / 180;